add town square; fix some room joining logic; firecracker item implemented; improved room player structures; improve inventory handling; update xml

+++ game-server/src/main/java/com/jftse/emulator/server/core/handler/lobby/room/RoomKickPlayerRequestPacketHandler.java
@@ -51,7 +51,7 @@ public class RoomKickPlayerRequestPacketHandler extends AbstractPacketHandler {
 
                 if (client != null) {
                     Packet answerPacket = new Packet(PacketOperations.S2CRoomLeaveAnswer);
-                    answerPacket.write(0);
+                    answerPacket.write((short) 0);
                     client.getConnection().sendTCP(answerPacket);
 
                     GameManager.getInstance().handleRoomPlayerChanges(client.getConnection(), true);
